+++ src/error_stack/result.rs
@@ -2,16 +2,16 @@ use core::fmt;
 
 use crate::{Context, Report};
 
-/// [`Result`](std::result::Result)`<T, `[`Report<C>`](Report)`>`
+/// [`std::result::Result`]`<T, `[`Report<C>`](Report)`>`
 ///
 /// A reasonable return type to use throughout an application.
 ///
-/// The `Result` type can be used with one or two parameters, where the first parameter represents
+/// The `BigResult` type can be used with one or two parameters, where the first parameter represents
 /// the [`Ok`] arm and the second parameter `Context` is used as in [`Report<C>`].
 ///
 /// # Examples
 ///
-/// `Result` can also be used in `fn main()`:
+/// `BigResult` can also be used in `fn main()`:
 ///
 /// ```rust
 /// # fn has_permission(_: (), _: ()) -> bool { true }
@@ -22,9 +22,9 @@ use crate::{Context, Report};
 /// #    fn fmt(&self, _: &mut std::fmt::Formatter<'_>) -> std::fmt::Result { Ok(()) }
 /// # }
 /// # impl bigerror::Context for AccessError {}
-/// use bigerror::{ensure, error_stack::Result};
+/// use bigerror::{ensure, BigResult};
 ///
-/// fn main() -> Result<(), AccessError> {
+/// fn main() -> BigResult<(), AccessError> {
 ///     let user = get_user()?;
 ///     let resource = get_resource()?;
 ///
@@ -38,25 +38,25 @@ use crate::{Context, Report};
 ///     # }; Ok(())
 /// }
 /// ```
-pub type Result<T, C> = core::result::Result<T, Report<C>>;
+pub type BigResult<T, C> = core::result::Result<T, Report<C>>;
 
-/// Extension trait for [`Result`][core::result::Result] to provide context information on
+/// Extension trait for [`core::result::Result`] to provide context information on
 /// [`Report`]s.
 pub trait ResultExt {
-    /// The [`Context`] type of the [`Result`].
+    /// The [`Context`] type of the [`BigResult`].
     type Context: Context;
 
-    /// Type of the [`Ok`] value in the [`Result`]
+    /// Type of the [`Ok`] value in the [`BigResult`]
     type Ok;
 
-    /// Adds a new attachment to the [`Report`] inside the [`Result`].
+    /// Adds a new attachment to the [`Report`] inside the [`BigResult`].
     ///
     /// Applies [`Report::attach`] on the [`Err`] variant, refer to it for more information.
     fn attach<A>(self, attachment: A) -> core::result::Result<Self::Ok, Report<Self::Context>>
     where
         A: Send + Sync + 'static;
 
-    /// Lazily adds a new attachment to the [`Report`] inside the [`Result`].
+    /// Lazily adds a new attachment to the [`Report`] inside the [`BigResult`].
     ///
     /// Applies [`Report::attach`] on the [`Err`] variant, refer to it for more information.
     fn attach_lazy<A, F>(
@@ -67,7 +67,7 @@ pub trait ResultExt {
         A: Send + Sync + 'static,
         F: FnOnce() -> A;
 
-    /// Adds a new printable attachment to the [`Report`] inside the [`Result`].
+    /// Adds a new printable attachment to the [`Report`] inside the [`BigResult`].
     ///
     /// Applies [`Report::attach_printable`] on the [`Err`] variant, refer to it for more
     /// information.
@@ -78,7 +78,7 @@ pub trait ResultExt {
     where
         A: fmt::Display + fmt::Debug + Send + Sync + 'static;
 
-    /// Lazily adds a new printable attachment to the [`Report`] inside the [`Result`].
+    /// Lazily adds a new printable attachment to the [`Report`] inside the [`BigResult`].
     ///
     /// Applies [`Report::attach_printable`] on the [`Err`] variant, refer to it for more
     /// information.
@@ -90,14 +90,14 @@ pub trait ResultExt {
         A: fmt::Display + fmt::Debug + Send + Sync + 'static,
         F: FnOnce() -> A;
 
-    /// Changes the context of the [`Report`] inside the [`Result`].
+    /// Changes the context of the [`Report`] inside the [`BigResult`].
     ///
     /// Applies [`Report::change_context`] on the [`Err`] variant, refer to it for more information.
     fn change_context<C>(self, context: C) -> core::result::Result<Self::Ok, Report<C>>
     where
         C: Context;
 
-    /// Lazily changes the context of the [`Report`] inside the [`Result`].
+    /// Lazily changes the context of the [`Report`] inside the [`BigResult`].
     ///
     /// Applies [`Report::change_context`] on the [`Err`] variant, refer to it for more information.
     fn change_context_lazy<C, F>(self, context: F) -> core::result::Result<Self::Ok, Report<C>>
@@ -114,7 +114,7 @@ where
     type Ok = T;
 
     #[track_caller]
-    fn attach<A>(self, attachment: A) -> Result<T, C>
+    fn attach<A>(self, attachment: A) -> BigResult<T, C>
     where
         A: Send + Sync + 'static,
     {
@@ -125,7 +125,7 @@ where
     }
 
     #[track_caller]
-    fn attach_lazy<A, F>(self, attachment: F) -> Result<T, C>
+    fn attach_lazy<A, F>(self, attachment: F) -> BigResult<T, C>
     where
         A: Send + Sync + 'static,
         F: FnOnce() -> A,
@@ -137,7 +137,7 @@ where
     }
 
     #[track_caller]
-    fn attach_printable<A>(self, attachment: A) -> Result<T, C>
+    fn attach_printable<A>(self, attachment: A) -> BigResult<T, C>
     where
         A: fmt::Display + fmt::Debug + Send + Sync + 'static,
     {
@@ -148,7 +148,7 @@ where
     }
 
     #[track_caller]
-    fn attach_printable_lazy<A, F>(self, attachment: F) -> Result<T, C>
+    fn attach_printable_lazy<A, F>(self, attachment: F) -> BigResult<T, C>
     where
         A: fmt::Display + fmt::Debug + Send + Sync + 'static,
         F: FnOnce() -> A,
@@ -160,7 +160,7 @@ where
     }
 
     #[track_caller]
-    fn change_context<C2>(self, context: C2) -> Result<T, C2>
+    fn change_context<C2>(self, context: C2) -> BigResult<T, C2>
     where
         C2: Context,
     {
@@ -171,7 +171,7 @@ where
     }
 
     #[track_caller]
-    fn change_context_lazy<C2, F>(self, context: F) -> Result<T, C2>
+    fn change_context_lazy<C2, F>(self, context: F) -> BigResult<T, C2>
     where
         C2: Context,
         F: FnOnce() -> C2,
@@ -183,7 +183,7 @@ where
     }
 }
 
-impl<T, C> ResultExt for Result<T, C>
+impl<T, C> ResultExt for BigResult<T, C>
 where
     C: Context,
 {
@@ -241,7 +241,7 @@ where
     }
 
     #[track_caller]
-    fn change_context<C2>(self, context: C2) -> Result<T, C2>
+    fn change_context<C2>(self, context: C2) -> BigResult<T, C2>
     where
         C2: Context,
     {
@@ -253,7 +253,7 @@ where
     }
 
     #[track_caller]
-    fn change_context_lazy<C2, F>(self, context: F) -> Result<T, C2>
+    fn change_context_lazy<C2, F>(self, context: F) -> BigResult<T, C2>
     where
         C2: Context,
         F: FnOnce() -> C2,