+++ tests/test_debug.rs
@@ -70,12 +70,12 @@ fn snap_suffix() -> String {
 #[cfg(feature = "spantrace")]
 pub fn create_report() -> Report<RootError> {
     #[tracing::instrument]
-    fn func_b() -> error_stack::Result<(), RootError> {
+    fn func_b() -> error_stack::BigResult<(), RootError> {
         create_error()
     }
 
     #[tracing::instrument]
-    fn func_a() -> error_stack::Result<(), RootError> {
+    fn func_a() -> error_stack::BigResult<(), RootError> {
         func_b()
     }